fixed jackpot improperly ending
